backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
anthropic
deepgram-sdk
python-multipart
python-dotenv
alembic
